+++ package.json
@@ -41,6 +41,7 @@
     "react-hot-toast": "^2.4.0",
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
+    "socket.io-client": "^4.5.3",
     "typescript": "~4.1.5",
     "uuid": "^9.0.0",
     "yup": "^0.32.11",
