+++ src/components/Chat/MessageList/style.module.scss
@@ -1,0 +1,18 @@
+@import 'src/styles/_color.scss';
+@import 'src/styles/_mixin.scss';
+
+.profile-active {
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  max-width: 758px;
+  @include FadeUp();
+  cursor: pointer;
+}
+.profile-hidden {
+  @include FadeDown();
+}
+
+.chat-content {
+  padding: 0 12px;
+}
