+++ src/components/Chat/index.ts
@@ -1,4 +1,5 @@
-export { default as ChatCard } from './ChatCard';
 export { default as ChatList } from './ChatList';
 export { default as Message } from './Message';
+export { default as MessageCard } from './MessageCard';
 export { default as MessageInput } from './MessageInput';
+export { default as MessageList } from './MessageList';
