+++ src/types/chat.ts
@@ -1,0 +1,9 @@
+export type Chat = {
+  sender: {
+    id: string;
+    name: string;
+    imagePath: string;
+  };
+  content: string;
+  date: string;
+};
